requirements.txt
starlette==0.37.2
uvicorn[standard]==0.30.1
redis==5.0.4
pydantic==2.7.1
PyYAML==6.0.1
orjson==3.10.3
httpx==0.27.0
typer==0.12.3
